Create F = ma
## Conclusion

In Lab 3: F = ma, we are introduced to Newton's second law of motion which states that the acceleration of an object is directly proportional to the net force and inversely proportional to its mass. Given the equation F = ma, F represents force, m represents mass and a represents acceleration. In this lab we are testing out objects with different masses and applied forces to see how these factors affect acceleration of an object, velocity of the object and its position in motion. The relationship between position (p(t)), velocity (v(t)) and acceleration (a(t)) is that position is the displacement of an object from one inital point to point 1 relative to time. From this information we can take the derivative of position (p(t)) and find velocity as mentioned in Lab 1 the equation for velocity is v(t) = Displacement(dp) / Time Elapsed (dt). To find acceleration, the rate of change of velocity, you take the derivative of velocity to form this equation a(t) = dv/dt, which is the change in velocity relative to the time elapsed. These are related because to change the objects velocity, we must apply a force, which will cause a change in acceleration. This change is depicted in the graphs within our lab. 

In this lab we tested different masses and different values of force to an object and observed the changed in position, velocity, and acceleration. We also created a graph showing how the derivative, in this case, acceleration is identical and overlies the same marks on the first graph which represents force. This is because of the relationship between force and acceleration. As mentioned above, the acceleration of an object is directly proportional to its net force. In all 6 blocks of this lab we have tested out different variables and plotted 4 graphs depicting this information and we will discuss the differences. 

In blocks 1-2 we simply plotted the labs given values for position and velocity, in block 3, we then added a force of 2 from point 0.62s to 0.88s and a force of 30 from 0.22s to 0.38s, with object mass of 1.0. These graph shows the position of the object moving from 0-0.2s at a slight upward on a slightly parabolic line, from 0.38 s you can see that the parabola becomes more distinct and on a steeper positive slope, this indicated the force of 30, which represent a big kick, applied to the object. In this example, the smaller "kick" from 0.62s to 0.88s is there are very small changes because the net force applied is 2 which is not very strong as represented in the graphs. We are able to see in the velocity graph, where the linear slope is within that time frame it is not very steep and the graph representing the force, the horizontal line is not too far above the line where there is zero force applied.